+++ src/cache.ts
@@ -17,7 +17,6 @@ interface Stat {
 export class LruCache<K, C = Uint8Array> {
   public stat: Stat
   private cache: Map<K, C> = new Map()
-  private rareKeys?: Set<K>
 
   constructor(public name: string, public size = 30) {
     caches[name] = this
@@ -39,24 +38,6 @@ export class LruCache<K, C = Uint8Array> {
       return val
     }
 
-    // Prevent rare keys
-    if (this.rareKeys) {
-      // If this key appears first time, we keep record of it.
-      if (!this.rareKeys.has(key)) {
-        this.rareKeys.add(key)
-        return cb(key) as C
-      }
-
-      // If this key appears second time, we can cache it.
-      this.rareKeys.delete(key)
-
-      // Evict
-      if (this.rareKeys.size >= 200) {
-        this.stat.rare += this.rareKeys.size
-        this.rareKeys.clear()
-      }
-    }
-
     // Cache the value
     val = cb(key) as C
     this.cache.set(key, val)
@@ -72,9 +53,6 @@ export class LruCache<K, C = Uint8Array> {
   }
 
   noRareKeys() {
-    if (!this.rareKeys) {
-      this.rareKeys = new Set()
-    }
     return this
   }
 }